refactor: 공통된 생성 메서드를 fixture 로 분리
